Remove duplicated classes
