--- Vorp.Core.Client/Commands/CommandContext.cs
@@ -1,10 +1,0 @@
-namespace Vorp.Core.Client.Commands
-{
-    public abstract class CommandContext
-    {
-        public abstract string[] Aliases { get; set; }
-        public abstract string Title { get; set; }
-        public abstract bool IsRestricted { get; set; }
-        public abstract List<string> RequiredRoles { get; set; }
-    }
-}
