--- Vorp.Core.Client/Commands/CommandInfo.cs
@@ -1,14 +1,0 @@
-namespace Vorp.Core.Client.Commands
-{
-    public class CommandInfo : Attribute
-    {
-        public string[] Aliases { get; set; }
-        public string Description { get; set; }
-
-        public CommandInfo(string[] aliases, string description)
-        {
-            Aliases = aliases;
-            Description = description;
-        }
-    }
-}
